+++ src/al/kernel.h
@@ -7,18 +7,18 @@
 namespace ca {
 namespace kernel {
 
-const double norm_range = 4.0;
-
-enum Type { GAME_OF_LIFE, MOORE, FULL };
+enum Type { GAME_OF_LIFE, MOORE, FULL, EXPONENTIAL };
 
 double Kernel(const Grid& grid, const int& x, const int& y, const int& radius, const Type& type);
-double Kernel(const double& value, const int& x_diff, const int y_diff, const Type& type);
-double Rectangular(const double& x_diff, const double& y_diff, const double& min, const double& max);
+double Kernel(const double& value, const int& x_diff, const int y_diff, const double &radius, const Type& type);
+double Range(const int& x_diff, const int y_diff, const double& radius);
+double Rectangular(const double& x_diff, const double& y_diff, const double& radius, const double& min, const double& max);
+double Exponential(const double& x_diff, const double& y_diff, const double& radius, const double& alpha = 4.0);
 
 // implementation
-double GameOfLife(const double& x_diff, const double& y_diff);
-double Moore(const double& x_diff, const double& y_diff);
-double Full(const double& x_diff, const double& y_diff);
+double GameOfLife(const double& x_diff, const double& y_diff, const double& radius);
+double Moore(const double& x_diff, const double& y_diff, const double& radius);
+double Full(const double& x_diff, const double& y_diff, const double& radius);
 
 }  // namespace kernel
 }  // namespace ca