+++ src/al/lenia.h
@@ -1,6 +1,10 @@
 #pragma once
 
+#include <memory>
+
 #include "../grid/grid.h"
+#include "../kernel/exponential.h"
+#include "../kernel/game_of_life.h"
 #include "growth_mapping.h"
 #include "kernel.h"
 
@@ -12,17 +16,28 @@ class Lenia {
         uint x = 1;
         uint y = 1;
         uint radius = 1;
-        kernel::Type type = kernel::GAME_OF_LIFE;
+        kernel::Type kernl_type = kernel::GAME_OF_LIFE;
+        growth_mapping::Type growth_type = growth_mapping::RECTANGULAR;
         double dt = 1.0;
         double mu = 0.0;
         double sigma = 1.0;
 
         void GameOfLife() {
             radius = 1;
-            type = kernel::GAME_OF_LIFE;
+            kernl_type = kernel::GAME_OF_LIFE;
+            growth_type = growth_mapping::RECTANGULAR;
             dt = 1.0;
-            mu = 3.0;
-            sigma = 0.5;
+            mu = 0.35;
+            sigma = 0.07;
+        }
+
+        void Lenia() {
+            radius = 12;
+            kernl_type = kernel::EXPONENTIAL;
+            growth_type = growth_mapping::EXPONENTIAL;
+            dt = 0.1;
+            mu = 0.15;
+            sigma = 0.016;
         }
     };
 
@@ -48,6 +63,7 @@ class Lenia {
     Config config_;
     Grid grid1_;
     Grid grid2_;
+    std::shared_ptr<Kernel> kernel_;
 };
 
 }  // namespace ca