+++ src/gui/gui.cc
@@ -7,10 +7,10 @@ Random GUI::random_;
 uint GUI::counter_;
 
 GUI::GUI() {
-    const uint size = 50;
+    const uint size = 100;
     Lenia::Config config;
 
-    config.GameOfLife();
+    config.Lenia();
 
     config.x = size;
     config.y = size;
@@ -19,19 +19,45 @@ GUI::GUI() {
 
     lenia_.SetConfig(config);
 
-    for (size_t x = 0; x < lenia_.size_x(); x++) {
-        for (size_t y = 0; y < lenia_.size_y(); y++) {
-            lenia_.value(x, y) = std::round(random_.RandomNumber());
+    for (size_t x = 0; x < lenia_.size_x() - 0; x++) {
+        for (size_t y = 0; y < lenia_.size_y() - 0; y++) {
+            lenia_.value(x, y) = 0.0;
+        }
+    }
+
+    for (size_t x = 30; x < 70; x++) {
+        for (size_t y = 30; y < 70; y++) {
+            lenia_.value(x, y) = random_.RandomNumber();
         }
     }
 }
 
 void GUI::DrawGrid(void) {
+    const double X = glutGet(GLUT_WINDOW_WIDTH);
+    const double Y = glutGet(GLUT_WINDOW_HEIGHT);
+    const double dx = X > Y ? (Y / X) * 2.0 / lenia_.size_x() : 2.0 / lenia_.size_x();
+    const double dy = Y > X ? (X / Y) * 2.0 / lenia_.size_y() : 2.0 / lenia_.size_y();
+
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
     for (size_t x = 0; x < lenia_.size_x(); x++) {
         for (size_t y = 0; y < lenia_.size_y(); y++) {
-            DrawCell(x, y);
+            const double color1 = gui_algorithms::Red(lenia_.value(x, y));
+            const double color2 = gui_algorithms::Green(lenia_.value(x, y));
+            const double color3 = gui_algorithms::Blue(lenia_.value(x, y));
+            const double x__ = x * dx - 1.0;
+            const double y__ = y * dy - 1.0;
+
+            glBegin(GL_POLYGON);
+            {  // GL_POLYGON GL_LINE_LOOP
+                glColor3f(color1, color2, color3);
+
+                glVertex3f(x__, y__, 0.0);
+                glVertex3f(x__ + dx, y__, 0.0);
+                glVertex3f(x__ + dx, y__ + dy, 0.0);
+                glVertex3f(x__, y__ + dy, 0.0);
+            }
+            glEnd();
         }
     }
 
@@ -39,52 +65,33 @@ void GUI::DrawGrid(void) {
     glutSwapBuffers();
 }
 
-void GUI::DrawCell(int x, int y) {
-    const double dx = 0.9 / lenia_.size_x();
-    const double width = glutGet(GLUT_WINDOW_WIDTH);
-    const double height = glutGet(GLUT_WINDOW_HEIGHT);
-    const double dy = dx * width / height;
-    const double color1 = 255;
-    const double color2 = 0 + !lenia_.value(x, y) * 255;
-    const double color3 = 0 + !lenia_.value(x, y) * 255;
-
-    const double x__ = dx * (2 * x) + dx - 0.9;
-    const double y__ = dy * (2 * y) + dy - 0.9;
-
-    glBegin(GL_POLYGON);
-    {  // GL_POLYGON GL_LINE_LOOP
-        glColor3f(color1, color2, color3);
-
-        glVertex3f(x__ - dx, y__ - dy, 0.0);
-        glVertex3f(x__ + dx, y__ - dy, 0.0);
-        glVertex3f(x__ + dx, y__ + dy, 0.0);
-        glVertex3f(x__ - dx, y__ + dy, 0.0);
-    }
-    glEnd();
-
-    glBegin(GL_LINE_LOOP);
-    {  // GL_POLYGON GL_LINE_LOOP
-        glColor3f(0, 0, 0);
-
-        glVertex3f(x__ - dx, y__ - dy, 0.0);
-        glVertex3f(x__ + dx, y__ - dy, 0.0);
-        glVertex3f(x__ + dx, y__ + dy, 0.0);
-        glVertex3f(x__ - dx, y__ + dy, 0.0);
-    }
-    glEnd();
-}
-
 void GUI::SpecialFunc(int key, int x, int y) { std::cout << key << std::endl; }
 
 void GUI::KeyboardFunc(unsigned char key, int x, int y) {
     switch (key) {
         case 'r': {
-            /*for (uint i = 0; i < grid1_.size(); i++) {
-                grid1_.value(i) = random_.RandomNumber();
-            }*/
+            for (size_t x = 0; x < lenia_.size_x(); x++) {
+                for (size_t y = 0; y < lenia_.size_y(); y++) {
+                    lenia_.value(x, y) = random_.RandomNumber();
+                }
+            }
 
-            glutPostRedisplay();
+            // glutPostRedisplay();
 
+            break;
+        }
+        case 'o': {
+            for (size_t x = 0; x < lenia_.size_x() - 0; x++) {
+                for (size_t y = 0; y < lenia_.size_y() - 0; y++) {
+                    lenia_.value(x, y) = 0.0;
+                }
+            }
+
+            for (size_t x = 30; x < 70; x++) {
+                for (size_t y = 30; y < 70; y++) {
+                    lenia_.value(x, y) = random_.RandomNumber();
+                }
+            }
             break;
         }
     }