+++ test/unit_tests/al/kernel.h
@@ -12,10 +12,13 @@ TEST(kernel, Rectangular) {
     const uint y = 3;
     const uint x_centre = 1;
     const uint y_centre = 1;
+    const uint radius = 1;
     Grid grid(x, y);
     double value = 0.0;
     double tot_value = 0.0;
     double count = 0.0;
+    const double min = 0.0;
+    const double max = 1.42;
 
     for (size_t i = 0; i < grid.size_x(); i++) {
         for (size_t j = 0; j < grid.size_y(); j++) {
@@ -25,19 +28,41 @@ TEST(kernel, Rectangular) {
         }
     }
 
-    count += grid.value(x_centre - 1, y_centre - 1) * Rectangular(1.0, 1.0, 0.0, 0.75);
-    count += grid.value(x_centre - 1, y_centre + 0) * Rectangular(1.0, 0.0, 0.0, 0.75);
-    count += grid.value(x_centre - 1, y_centre + 1) * Rectangular(1.0, 1.0, 0.0, 0.75);
+    count += grid.value(x_centre - 1, y_centre - 1) * Rectangular(1.0, 1.0, radius, min, max);
+    count += grid.value(x_centre - 1, y_centre + 0) * Rectangular(1.0, 0.0, radius, min, max);
+    count += grid.value(x_centre - 1, y_centre + 1) * Rectangular(1.0, 1.0, radius, min, max);
 
-    count += grid.value(x_centre - 0, y_centre - 1) * Rectangular(0.0, 1.0, 0.0, 0.75);
-    count += grid.value(x_centre - 0, y_centre + 0) * Rectangular(0.0, 0.0, 0.0, 0.75);
-    count += grid.value(x_centre - 0, y_centre + 1) * Rectangular(0.0, 1.0, 0.0, 0.75);
+    count += grid.value(x_centre - 0, y_centre - 1) * Rectangular(0.0, 1.0, radius, min, max);
+    count += grid.value(x_centre - 0, y_centre + 0) * Rectangular(0.0, 0.0, radius, min, max);
+    count += grid.value(x_centre - 0, y_centre + 1) * Rectangular(0.0, 1.0, radius, min, max);
 
-    count += grid.value(x_centre + 1, y_centre - 1) * Rectangular(0.0, 1.0, 0.0, 0.75);
-    count += grid.value(x_centre + 1, y_centre + 0) * Rectangular(0.0, 0.0, 0.0, 0.75);
-    count += grid.value(x_centre + 1, y_centre + 1) * Rectangular(0.0, 1.0, 0.0, 0.75);
+    count += grid.value(x_centre + 1, y_centre - 1) * Rectangular(0.0, 1.0, radius, min, max);
+    count += grid.value(x_centre + 1, y_centre + 0) * Rectangular(0.0, 0.0, radius, min, max);
+    count += grid.value(x_centre + 1, y_centre + 1) * Rectangular(0.0, 1.0, radius, min, max);
 
     ASSERT_DOUBLE_EQ(tot_value, count);
+}
+
+TEST(kernel, Exponential) {
+    const uint x = 3;
+    const uint y = 3;
+    const uint x_centre = 1;
+    const uint y_centre = 1;
+    const uint radius = 2;
+    Grid grid(x, y);
+    double value = 137.11;
+
+    grid.value(x_centre, y_centre) = value;
+    grid.value(x_centre + 2, y_centre) = value;
+    grid.value(x_centre + 1, y_centre) = value;
+
+    double result1 = grid.value(x_centre, y_centre) * Exponential(0.0, 0.0, radius);
+    double result2 = grid.value(x_centre + 2, y_centre) * Exponential(2.0, 0.0, radius);
+    double result3 = grid.value(x_centre + 1, y_centre) * Exponential(1.0, 0.0, radius);
+
+    ASSERT_DOUBLE_EQ(0.0, result1);
+    ASSERT_DOUBLE_EQ(0.0, result2);
+    ASSERT_DOUBLE_EQ(value, result3);
 }
 
 TEST(kernel, Moore) {
@@ -64,7 +89,7 @@ TEST(kernel, Full) {
     const uint y = 9;
     const uint x_centre = 4;
     const uint y_centre = 4;
-    const uint radius = 9;
+    const uint radius = 3;
     Grid grid(x, y);
 
     for (size_t i = 0; i < grid.size_x(); i++) {
@@ -94,7 +119,7 @@ TEST(kernel, GameOfLife) {
 
     const double count = Kernel(grid, x_centre, y_centre, radius, GAME_OF_LIFE);
 
-    ASSERT_DOUBLE_EQ(8.5, count);
+    ASSERT_DOUBLE_EQ(1.0, count);
 }
 
 TEST(kernel, GameOfLife_BeeHive) {
@@ -116,7 +141,7 @@ TEST(kernel, GameOfLife_BeeHive) {
 
     const double count = Kernel(grid, x_centre, y_centre, radius, GAME_OF_LIFE);
 
-    ASSERT_DOUBLE_EQ(1.0, count);
+    ASSERT_DOUBLE_EQ(1.0 / 8.5, count);
 }
 
 }  // namespace kernel